add fancy logger

+++ scripts/preprocessing.py
@@ -25,8 +25,6 @@ class RoadDataset(Dataset):
     """
 
     def __init__(self, image_paths, mask_paths, transform=None):
-        #super().__init__(image_paths, *args, **kwargs)
-
         # read images in
         self.images = [mpimg.imread(path) for path in image_paths]
         self.masks = [mpimg.imread(path) for path in mask_paths]
@@ -37,8 +35,8 @@ class RoadDataset(Dataset):
         trimap = np.array(Image.fromarray((self.masks[i] * 255).astype(np.uint8)).resize((512, 512)))
         mask = np.where(trimap > 128, 1, 0)
         
-        image = image.astype(np.float32)    # Previous division creates float64 by default, but augmentations can only handle one of [uint8, float32]
-        mask = mask.astype(np.uint8)        # Mask values are [0,1] so uint8 is more reasonable: np.uint32 -> np.uint8
+        image = image.astype(np.float32)    # augmentations can only handle one of [uint8, float32]
+        mask = mask.astype(np.uint8)        # values are {0,1} so uint8 is more reasonable: np.uint32 -> np.uint8
         
         if self.transform:
             # Apply same transformation to image and mask
